backend/requirements.txt
flask
opencv-python
mediapipe
numpy
ultralytics
reportlab

psycopg2-binary
